Redesign the partitioning strategy to simplify things
Instead of allowing zoltan to auto_migrate points, we hold onto them,
examine cells and add extra points which need to be transferred to the
lists. The points on cells which are split are kept if necessary and a
single point exchange takes place menaing we only need one big memory
allocation.

Cells can be transferred as before after intermediate point storafe is
cleared. Overall memory usage should be a little less.