Inflate bounds when resampling, add Bounds Translator support
The SPH resampler can enlarge the dataset bounds by adding
padding regions for sampling. This means we must inflate the region
bounds and also the data bounds within them, but without changing
the split planes that the partitioning chose.

Add support for this to the BoundsExtentTranslator (BET) and modify the
SPH resampler to create a new BET rather than modify the one passed in.