Replace array allocation which uses non-standard compiler extension
Arrays allocated using the syntax
int array[non_const_value];
are only valid during the scope of th block they are declared in
and one the function is left, the pointers are invalid.
This caused a segfault later.

The arrays are now allocated with new int[num] but are not deleted
this needs to be fixed.